+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Fixes
 - Log rotator task - daily, weekly, monthly and yearly schedule (issue #49)
+- Improved the command line handler to support arrow keys and history using up key
 
 ### Updates
 - Debug printed in server version in connection response if build has debug logs enabled